+++ odoo/local-src/schweizmobil_subs_bundle_ee/__manifest__.py
@@ -12,6 +12,12 @@
         'account_bank_statement_import_camt',
         'account_followup',
         'sale_subscription',
+        # external-src/odoo-enterprise-addons
+        'sale_subscription_to_close',
+        'sale_subscription_closing_delay',
+        'sale_subscription_invoice_cancel',
+        'sale_subscription_next_invoice_date_advance',
+        'sale_subscription_to_close_next_date_advance'
     ],
     'installable': True,
     'auto_install': False,